# server/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_yogesh4952_ebookstore_internal_cart_model.AddToCartPayload:
    properties:
      book_id:
        type: integer
      quantity:
        type: integer
    required:
    - book_id
    - quantity
    type: object
  github_com_yogesh4952_ebookstore_internal_order_models.OrderItemRequest:
    properties:
      book_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
      seller_id:
        type: integer
    required:
    - book_id
    - quantity
    - seller_id
    type: object
  github_com_yogesh4952_ebookstore_internal_order_models.PaymentMethod:
    enum:
    - COD
    - ESEWA
    type: string
    x-enum-varnames:
    - PayementCOD
    - PayementEsewa
  github_com_yogesh4952_ebookstore_internal_order_models.PlaceOrderPayload:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_order_models.OrderItemRequest'
        minItems: 1
        type: array
      payment_method:
        $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_order_models.PaymentMethod'
      user_address_id:
        type: integer
    required:
    - items
    - payment_method
    - user_address_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the EBook Store platform
  title: EBook Store API
  version: "1.0"
paths:
...
  /cart/add-to-cart:
    post:
      consumes:
      - application/json
      description: 'Add a quantity of a book to the authenticated user''s cart. Quantities
        are additive: adding the same book twice increments the stored count rather
        than replacing it. The user id is taken from the JWT, so a caller can only
        ever modify their own cart.'
      parameters:
      - description: Book and quantity to add
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_cart_model.AddToCartPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Book added to cart, containing success and message keys
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Book not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: User id on the request context had an unexpected type
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add a book to the cart
      tags:
      - cart
  /order/list-user-order:
    get:
      consumes:
      - application/json
      description: List every order belonging to the authenticated user, including
        their line items. The authenticated user id is taken from the JWT and cannot
        be overridden in the request.
      produces:
      - application/json
      responses:
        "202":
          description: Orders fetched, containing the list of orders under data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Orders could not be fetched
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: User id on the request context had an unexpected type
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List User order
      tags:
      - orders
  /order/place-order:
    post:
      consumes:
      - application/json
      description: Create a new order with items, payment method, and shipping address.
        For the `COD` payment method the order is created as PENDING. For the `ESEWA`
        method the response additionally carries a signed `esewa_payload` that the
        client must POST to eSewa's payment form; the order stays PENDING until the
        eSewa callback is confirmed at `/payment/success`. Prices are always recalculated
        server-side from the book records, so any price in the request is ignored.
      parameters:
      - description: Order data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_yogesh4952_ebookstore_internal_order_models.PlaceOrderPayload'
      produces:
      - application/json
      responses:
        "202":
          description: Order placed, containing the order under data. For ESEWA orders
            this also includes the signed esewa_payload
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body, or an unsupported payment method
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Address id or one of the book ids was not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: User id on the request context had an unexpected type, or the
            order could not be created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Place a new order
      tags:
      - orders
  /payment/failure:
    get:
      description: |-
        Browser-facing endpoint that eSewa redirects the customer to when they cancel at the payment page or the payment is declined. There is no payment to verify on a cancel, so the attempt is logged and the browser is redirected to the frontend. This endpoint is unauthenticated because the caller is the customer's browser arriving from eSewa.

        The order is deliberately left in the PENDING payment status: a cancelled checkout is not a dead order, and the customer is free to retry with the same transaction uuid.
      parameters:
      - description: Transaction identifier sent by eSewa (also accepted as transaction_uuid)
        in: query
        name: TransactionUuid
        type: string
      - description: Failure reason reported by eSewa
        in: query
        name: reason
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to FRONTEND_URL/payment-failure
          schema:
            type: string
      summary: Handle a cancelled or failed eSewa payment
      tags:
      - payment
  /payment/success:
    get:
      description: |-
        Browser-facing endpoint that eSewa redirects the customer to after a payment. It decodes the base64 `data` query parameter, verifies the callback signature, confirms the transaction against eSewa's status API, marks the order paid, and redirects the browser to the frontend. This endpoint is unauthenticated because the caller is the customer's browser arriving from eSewa, which cannot present a JWT.

        On success the response is a 302 redirect to `FRONTEND_URL/payment-success?orderId=<transaction_uuid>`. On a business failure it is a 302 redirect to `FRONTEND_URL/payment-failure?reason=<reason>`, where reason is one of `invalid_signature`, `order_not_found`, `order_not_payable`, `amount_mismatch`, `payment_incomplete`, `verification_unavailable`, or `verification_failed`.

        A 400 is returned only when the `data` parameter itself is missing or undecodable. Replaying this endpoint for an already-paid order is safe: it returns the same success redirect without re-running fulfillment.
      parameters:
      - description: Base64-encoded JSON callback signed by eSewa
        in: query
        name: data
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to FRONTEND_URL/payment-success with the order id
            as a query parameter
          schema:
            type: string
        "400":
          description: Missing or undecodable data parameter
          schema:
            additionalProperties: true
            type: object
      summary: Confirm an eSewa payment
      tags:
      - payment
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"